Reduce code duplication in template codegen and fix std::HashSet usage
- Extract shared event handler functions (build_legacy_event_handler,
  gen_legacy_event_on, gen_event_attr_on) into attributes.rs, eliminating
  identical implementations from svelte_body.rs, svelte_document.rs, and
  svelte_window.rs
- Extract binding getter/setter builders (build_binding_getter,
  build_binding_setter, build_binding_setter_silent) as pub(crate)
  functions, replacing duplicated closures across 3 files
- Replace std::collections::HashSet with FxHashSet in svelte_js

Net reduction: ~250 lines of duplicated code.
