+++ crates/svelte_js/Cargo.toml
@@ -13,6 +13,7 @@ oxc_semantic = { workspace = true }
 oxc_span = { workspace = true }
 oxc_syntax = { workspace = true }
 compact_str = { workspace = true }
+rustc-hash = { workspace = true }
 
 [dev-dependencies]
 pretty_assertions = { workspace = true }
